pyproject.toml
[tool.poetry]
authors = [
  "Kevin Amado <kamadorueda@gmail.com>"
]
description = "Open-Source APM (Application monitoring) project that offers you zero overhead wrappers for profiling your code execution flow"
license = "MIT"
name = "tracers"
packages = [
  { include = "tracers", from = "src" },
]
readme = "README.md"
repository = "https://github.com/kamadorueda/tracers"
version = "1.0.0"

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
prospector = "1.2.0"
python-dateutil = "2.8.1"
mypy = "0.770"

[build-system]
requires = [
  "poetry>=0.12"
]
build-backend = "poetry.masonry.api"
